+++ service/history/events/cache.go
@@ -25,11 +25,9 @@ package events
 import (
 	"time"
 
-	"github.com/uber/cadence/common/cache"
-
-	"github.com/uber/cadence/common"
-
 	"github.com/uber/cadence/.gen/go/shared"
+	"github.com/uber/cadence/common"
+	"github.com/uber/cadence/common/cache"
 	"github.com/uber/cadence/common/log"
 	"github.com/uber/cadence/common/log/tag"
 	"github.com/uber/cadence/common/metrics"
