+++ service/history/execution/mutable_state_task_refresher.go
@@ -25,8 +25,6 @@ package execution
 import (
 	"time"
 
-	"github.com/uber/cadence/service/history/events"
-
 	"github.com/uber/cadence/.gen/go/shared"
 	"github.com/uber/cadence/common"
 	"github.com/uber/cadence/common/cache"
@@ -34,6 +32,7 @@ import (
 	"github.com/uber/cadence/common/log"
 	"github.com/uber/cadence/common/persistence"
 	"github.com/uber/cadence/service/history/config"
+	"github.com/uber/cadence/service/history/events"
 )
 
 var emptyTasks = []persistence.Task{}
