+++ service/history/shard/contextTest.go
@@ -23,15 +23,14 @@ package shard
 import (
 	"time"
 
-	"github.com/uber/cadence/service/history/events"
-	"github.com/uber/cadence/service/history/resource"
-
 	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/mock"
 
 	"github.com/uber/cadence/common/metrics"
 	"github.com/uber/cadence/common/persistence"
 	"github.com/uber/cadence/service/history/config"
+	"github.com/uber/cadence/service/history/events"
+	"github.com/uber/cadence/service/history/resource"
 )
 
 // TestContext is a test implementation for shard Context interface
